# apps/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/payment-requests": {
      "post": {
        "description": "Crée une demande de paiement pour un client. Elle a une date d’expiration et peut être suivie jusqu’à réussite ou expiration.",
        "operationId": "PaymentRequestsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount",
                  "currency_code",
                  "expiry_date"
                ],
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "currency_code": {
                    "type": "string",
                    "enum": [
                      "XOF",
                      "USD",
                      "EUR"
                    ]
                  },
                  "description": {
                    "type": "string"
                  },
                  "customer_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "expiry_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "payment_reference": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              },
              "examples": {
                "default": {
                  "summary": "Demande avec expiration",
                  "value": {
                    "amount": 25000,
                    "currency_code": "XOF",
                    "description": "Honoraires conseil, mars",
                    "expiry_date": "2026-12-31T23:59:59.000Z",
                    "payment_reference": "INV-2026-042"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Demande de paiement créée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequestResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Entrée invalide ou erreur de validation"
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Créer une demande de paiement",
        "tags": [
          "Payment Requests"
        ]
      },
      "get": {
        "description": "Renvoie toutes les demandes de l'organisation avec pagination et filtres optionnels (statut, client).",
        "operationId": "PaymentRequestsController_findAll",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Décalage pour la pagination",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre maximal de résultats",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "description": "Filtrer par ID client",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtrer par statut",
            "schema": {
              "enum": [
                "pending",
                "completed",
                "failed",
                "expired"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée de demandes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PaymentRequestResponseDto"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "offset": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Lister les demandes de paiement",
        "tags": [
          "Payment Requests"
        ]
      }
    },
    "/payment-requests/{id}": {
      "get": {
        "description": "Renvoie le détail d’une demande (statut actuel, informations de paiement).",
        "operationId": "PaymentRequestsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID de la demande",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de la demande",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequestResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          },
          "404": {
            "description": "Demande introuvable ou accès refusé"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Obtenir une demande de paiement par ID",
        "tags": [
          "Payment Requests"
        ]
      }
    },
...
  },
  "info": {
    "title": "API lomi.",
    "description": "API de traitement des paiements pour les entreprises d'Afrique de l'Ouest francophone.",
    "version": "1.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.lomi.africa",
      "description": "Live"
    },
    {
      "url": "https://sandbox.api.lomi.africa",
      "description": "Test"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    },
    "schemas": {
...
      "PaymentRequestResponseDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 123
          },
          "created_at": {
            "type": "string",
            "example": "string"
          },
          "created_by": {
            "type": "string",
            "example": "string"
          },
          "currency_code": {
            "type": "string",
            "example": "string"
          },
          "customer_id": {
            "type": "string",
            "example": "string"
          },
          "description": {
            "type": "string",
            "example": "string"
          },
          "environment": {
            "type": "string",
            "example": "string"
          },
          "expiry_date": {
            "type": "string",
            "example": "string"
          },
          "organization_id": {
            "type": "string",
            "example": "string"
          },
          "payment_link": {
            "type": "string",
            "example": "string"
          },
          "payment_reference": {
            "type": "string",
            "example": "string"
          },
          "request_id": {
            "type": "string",
            "example": "string"
          },
          "spi_account_number": {
            "type": "string",
            "example": "string"
          },
          "spi_bulk_instruction_id": {
            "type": "string",
            "example": "string"
          },
          "spi_confirmation": {
            "type": "boolean",
            "example": true
          },
          "spi_date_envoi": {
            "type": "string",
            "example": "string"
          },
          "spi_date_irrevocabilite": {
            "type": "string",
            "example": "string"
          },
          "spi_date_limite_paiement": {
            "type": "string",
            "example": "string"
          },
          "spi_date_limite_reponse": {
            "type": "string",
            "example": "string"
          },
          "spi_date_rejet": {
            "type": "string",
            "example": "string"
          },
          "spi_debit_differe": {
            "type": "boolean",
            "example": true
          },
          "spi_end2end_id": {
            "type": "string",
            "example": "string"
          },
          "spi_payeur_alias": {
            "type": "string",
            "example": "string"
          },
          "spi_payeur_nom": {
            "type": "string",
            "example": "string"
          },
          "spi_payeur_pays": {
            "type": "string",
            "example": "string"
          },
          "spi_ref_doc_numero": {
            "type": "string",
            "example": "string"
          },
          "spi_remise_amount": {
            "type": "number",
            "example": 123
          },
          "spi_remise_rate": {
            "type": "number",
            "example": 123
          },
          "spi_tx_id": {
            "type": "string",
            "example": "string"
          },
          "status": {
            "type": "string",
            "example": "string"
          },
          "updated_at": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "amount",
          "created_at",
          "created_by",
          "currency_code",
          "customer_id",
          "description",
          "environment",
          "expiry_date",
          "organization_id",
          "payment_link",
          "payment_reference",
          "request_id",
          "spi_account_number",
          "spi_bulk_instruction_id",
          "spi_confirmation",
          "spi_date_envoi",
          "spi_date_irrevocabilite",
          "spi_date_limite_paiement",
          "spi_date_limite_reponse",
          "spi_date_rejet",
          "spi_debit_differe",
          "spi_end2end_id",
          "spi_payeur_alias",
          "spi_payeur_nom",
          "spi_payeur_pays",
          "spi_ref_doc_numero",
          "spi_remise_amount",
          "spi_remise_rate",
          "spi_tx_id",
          "status",
          "updated_at"
        ]
      },
...
    }
  }
}